+++ models/options/GlitchOptionSet.py
@@ -23,7 +23,7 @@ class GlitchOptionSet():
         self.whale_early = False
         self.sushiesless_toad_town_star_piece = False
         self.toad_town_sushie_glitch_gearless = False
-        self.toad_town_sushie_glitch = False
+        self.toad_town_sushie_glitch_onegear = False
         self.toad_town_sushie_glitch_full_gear = False
         self.jumpless_summit_climb = False