+++ models/options/OptionSet.py
@@ -591,8 +591,8 @@ class OptionSet:
             self.glitch_settings.sushiesless_toad_town_star_piece = options_dict.get("SushielessToadTownStarPiece")
         if "ToadTownSushieGlitchGearless" in options_dict:
             self.glitch_settings.toad_town_sushie_glitch_gearless = options_dict.get("ToadTownSushieGlitchGearless")
-        if "ToadTownSushieGlitch" in options_dict:
-            self.glitch_settings.toad_town_sushie_glitch = options_dict.get("ToadTownSushieGlitch")
+        if "ToadTownSushieGlitchOneGear" in options_dict:
+            self.glitch_settings.toad_town_sushie_glitch_onegear = options_dict.get("ToadTownSushieGlitchOneGear")
         if "ToadTownSushieGlitchFullGear" in options_dict:
             self.glitch_settings.toad_town_sushie_glitch_full_gear = options_dict.get("ToadTownSushieGlitchFullGear")
 
@@ -1399,7 +1399,7 @@ class OptionSet:
         basic_assert("WhaleEarly", bool)
         basic_assert("SushielessToadTownStarPiece", bool)
         basic_assert("ToadTownSushieGlitchGearless", bool)
-        basic_assert("ToadTownSushieGlitch", bool)
+        basic_assert("ToadTownSushieGlitchOneGear", bool)
         basic_assert("ToadTownSushieGlitchFullGear", bool)
 
         basic_assert("JumplessSummitClimb", bool)
@@ -1946,7 +1946,7 @@ class OptionSet:
         web_settings["WhaleEarly"]= self.glitch_settings.whale_early
         web_settings["SushielessToadTownStarPiece"] = self.glitch_settings.sushiesless_toad_town_star_piece
         web_settings["ToadTownSushieGlitchGearless"] = self.glitch_settings.toad_town_sushie_glitch_gearless
-        web_settings["ToadTownSushieGlitch"] = self.glitch_settings.toad_town_sushie_glitch
+        web_settings["ToadTownSushieGlitchOneGear"] = self.glitch_settings.toad_town_sushie_glitch_onegear
         web_settings["ToadTownSushieGlitchFullGear"] = self.glitch_settings.toad_town_sushie_glitch_full_gear
 
         # Glitches: Shooting Star Summit